Adding angular js support with ng boilerplate. Change in table creation strategy
under src/main/webapp
git clone https://github.com/ngbp/ngbp.git app

now to download the grunt dependencies which are listed in package.json
we have to do a npm install on the app directory
This will download the files to the node_modules

to download client side dependencies specified in bowerrc for the
application we do
bower install
this will download the js file to the vendor directory.

The build.config.js file will have the entry for all the files
downloaded into the vendor directory.